CLASS_ALIGNED_MVP_UI_DEMO/requirements.txt
streamlit
pypdf
python-docx
python-dotenv
google-genai